requirements.txt
moviepy
openai-whisper
click